% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_zoo.R
\name{add.zoo}
\alias{add.zoo}
\title{add.zoo}
\usage{
add.zoo(taxo)
}
\arguments{
\item{taxo}{taxonomic table generated with "add_taxo"}
}
\value{
A taxonomic table with also trophic levels.
}
\description{
Add trophic levels and OTU.
}
