% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopart_add-mod_zoo.R
\name{add_col}
\alias{add_col}
\title{add_col}
\usage{
add_col(df, column, col_name)
}
\arguments{
\item{df}{a tsv ecotaxa dataframe}

\item{column}{the colum to add}

\item{col_name}{the name to assign new column}
}
\description{
This function adds a column to the .tsv data frame
}
\author{
Alex Barth
}
