% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bin-taxa.R
\name{add_count_zeros}
\alias{add_count_zeros}
\title{Inside function for force bins option}
\usage{
add_count_zeros(tdf, taxa_names)
}
\arguments{
\item{tdf}{the split dataframe}

\item{taxa_names}{all possible taxa names}
}
\description{
Inside function for force bins option
}
