% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsvTrim.R
\name{add_custom_trim}
\alias{add_custom_trim}
\title{Call this function to edit vectorSelect}
\usage{
add_custom_trim()
}
\description{
BE CAREFUL. this will edit the package to your local version of the package
This makes your script less sharable but easiy if you are doing repeated
work in house. If you reinstall package it will restore to the default
}
