% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopart_par_conc.R
\name{add_par_zeros}
\alias{add_par_zeros}
\title{Add particle zeros}
\usage{
add_par_zeros(par_sum, full_seq)
}
\arguments{
\item{par_sum}{a par_sum object from calc_par_conc}

\item{full_seq}{the full sequence}
}
\description{
Just an inside function I wrote at 2am in the middle of the pacific
}
