% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{add_range}
\alias{add_range}
\title{add_range}
\usage{
add_range(number, width)
}
\arguments{
\item{number}{center number}

\item{width}{range on both sided}
}
\description{
copied from helper box
}
