% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_formatting.R
\name{add_zero_rows}
\alias{add_zero_rows}
\title{add_zero_rows
for bin_by, if there are no plankton in a bin it leaves it empty,
This fills those rows}
\usage{
add_zero_rows(df, ref_range)
}
\arguments{
\item{df}{a bin_by_df product}

\item{ref_range}{the reference range of data}
}
\description{
add_zero_rows
for bin_by, if there are no plankton in a bin it leaves it empty,
This fills those rows
}
