% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bin-helpers.R
\name{add_zeros}
\alias{add_zeros}
\title{add_zeros}
\usage{
add_zeros(df, col_name)
}
\arguments{
\item{df}{a data frame - should be in tidy format from bin_taxa}

\item{col_name}{a column to add 0s for}

\item{cat_col}{the category column}
}
\description{
This function adds 0's for the depth bins which don't have any present taxa
Will return a new data frame
}
\author{
Alex Barth
}
