% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopart_add_zoo.R
\name{add_zoo}
\alias{add_zoo}
\title{Add column to all zooplankton files}
\usage{
add_zoo(ecopart_obj, func, col_name, ...)
}
\arguments{
\item{ecopart_obj}{an ecopart object list}

\item{func}{function to apply (must take df and return vector) see ellps_vol}

\item{col_name}{name to assign to new column}

\item{...}{to pass to \code{lapply()} if needed}
}
\description{
This function will add a column to all zooplankton files in an ecopart_object
It returns a new ecopart object
}
\author{
Alex Barth
}
