% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopart_all_taxa.R
\name{all_taxa}
\alias{all_taxa}
\title{Get all taxa}
\usage{
all_taxa(ecopart_obj)
}
\arguments{
\item{ecopart_obj}{an ecopart object}
}
\description{
retrieve the names of all taxa in the ecopart_obj
}
\examples{
#getting all taxa
ecopart_example |> all_taxa()

#get unique taxa names
ecopart_example |>
  all_taxa() |>
  unique()

#get proportions
ecopart_example |> 
  all_taxa() |> 
  table() |> 
  prop.table()
}
