% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_helper-functions.R
\name{any_in}
\alias{any_in}
\title{Any In}
\usage{
any_in(checkVect, refVect)
}
\arguments{
\item{checkVect}{the input vector}

\item{refVect}{the reference vector}
}
\description{
Simply a wrapper for \%in\% to check if there are any
}
\examples{
any_in(c(0,4,3,1), c(1,7,6,8)) #returns TRUE
any_in(c(0,4,3), c(1,7,8,6)) #return FALSE

}
