% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_obj-volume.R
\name{area_to_esd}
\alias{area_to_esd}
\title{Convert Area to ESD}
\usage{
area_to_esd(area, pixel_mm)
}
\arguments{
\item{area}{the area in pixels}

\item{pixel_mm}{the pixel to mm conversion (not squared)}
}
\description{
If only the area of pixels are given you can convert it
to the ESD
}
