% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_init-functions.R
\name{as_ecopart_obj}
\alias{as_ecopart_obj}
\title{Set classes}
\usage{
as_ecopart_obj(obj = NULL, par_files = NULL, zoo_files = NULL, meta = NULL)
}
\arguments{
\item{obj}{the object}

\item{par_files}{the par files}

\item{zoo_files}{the zoo_files}

\item{meta}{the meta files}
}
\description{
Will set the class as an ecopart object
}
