% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_class_checks.R
\name{assign_etx_class}
\alias{assign_etx_class}
\title{Assign an etx class}
\usage{
assign_etx_class(input, add_class)
}
\arguments{
\item{input}{the input of object to assign a class}

\item{add_class}{the class to assign (can be a character vector)}
}
\description{
Primarily for looping in lapply
}
\author{
Alex Barth
}
