% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_size-spectra-tools.R
\name{assign_spectra_bins}
\alias{assign_spectra_bins}
\title{Make size bins in a cut style vector}
\usage{
assign_spectra_bins(
  size_vect,
  method,
  width = NULL,
  base = 2,
  min_size = NULL,
  max_size = NULL
)
}
\arguments{
\item{size_vect}{the vector to assign classes for}

\item{method}{log or width}

\item{width}{if width, need to specify the width}

\item{base}{if log method, the base of the log}

\item{min_size}{minimum size}

\item{max_size}{maximum size}
}
\description{
Runs for a single bin group
}
