% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopart_average_casts.R
\name{avg_casts}
\alias{avg_casts}
\title{Inside averaging}
\usage{
avg_casts(zoo_conc_list)
}
\arguments{
\item{zoo_conc_list}{a zoo conc list}
}
\description{
Takes a zoo_conc_list object and averages it
}
