% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binByTools.R
\name{bin_by}
\alias{bin_by}
\title{bin_by - a function to make easy cuts by default}
\usage{
bin_by(df, method, custom_range = NULL, equal_step = 10, equal_number = 100)
}
\arguments{
\item{df}{the vector to include to be cut. Should be an ecotaxa tsv export}

\item{method}{"Equal Space", "Custom", "Equal Number" "Zooscan"}

\item{custom_range}{vector of ranges}

\item{equal_step}{create equal steps of this size}

\item{equal_number}{create this number of ranges}
}
\description{
This fuction will return a vector for which depth bins for each observation
}
