% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binByTools.R
\name{bin_by_df}
\alias{bin_by_df}
\title{bin_by_df - returns a data frame for counts by bin}
\usage{
bin_by_df(df, method, custom_range, equal_step = 10, equal_number = 100)
}
\arguments{
\item{df}{the vector to include to be cut. Should be an ecotaxa tsv export}

\item{method}{"Equal Space", "Custom", "Equal Number" "Zooscan"}

\item{custom_range}{vector of ranges}

\item{equal_step}{create equal steps of this size}

\item{equal_number}{create this number of ranges}
}
\description{
will count number of observations for each bin then return a count dataframe
}
