% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bin-helpers.R
\name{bin_format_check}
\alias{bin_format_check}
\title{bin_format_check
returns true if all character of the vector match the '(#,#]' format}
\usage{
bin_format_check(input)
}
\arguments{
\item{input}{input from higher function}
}
\description{
bin_format_check
returns true if all character of the vector match the '(#,#]' format
}
\author{
Alex Barth
}
