% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bin-helpers.R
\name{bin_format_depth}
\alias{bin_format_depth}
\title{Legacy in interior for bin format}
\usage{
bin_format_depth(df)
}
\arguments{
\item{df}{the data frame}
}
\description{
same as general but will return formatted columns with _d
}
