% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bin-helpers.R
\name{bin_format_general}
\alias{bin_format_general}
\title{Interior for bin format}
\usage{
bin_format_general(df, col)
}
\arguments{
\item{df}{the data frame}

\item{col}{character name of column to format}
}
\description{
Generalized and will only provide min/max/mp columns
}
