% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopart_plot_par.R
\name{bin_par_plot}
\alias{bin_par_plot}
\title{Plot a size-binned particle plot}
\usage{
bin_par_plot(par_conc_df, ...)
}
\arguments{
\item{par_conc_df}{a particle conc df}

\item{...}{arguments to pass to \code{geom_smooth} can set gam or loess, span, etc}
}
\description{
Plot a size-binned particle plot
}
