% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bin-taxa.R
\name{bin_taxa}
\alias{bin_taxa}
\title{Bin observations by taxa according to some common factor
This version allows for summing over some factor like dry mass or volume}
\usage{
bin_taxa(
  df,
  depth_breaks,
  zooscan = F,
  cat_col = NULL,
  func_col = "taxo_name",
  func = length,
  max_d = NULL,
  force_bins = F
)
}
\arguments{
\item{df}{a data frame should have a taxo_name, name, or object_annotation_category column}

\item{depth_breaks}{the limits of what bins should be made}

\item{zooscan}{if the df is a zooscan export set to true, default false}

\item{cat_col}{a category other than taxa to bin by}

\item{func_col}{a single character vector of the value to apply function to; esd, biomass, drymass, etc}

\item{func}{the function to apply}

\item{max_d}{the maximum possible depth for a bin}

\item{force_bins}{set to true if you want to ensure observations for all 0's}
}
\description{
It will sum for each individual in a depth bin by taxa
}
\author{
Alex Barth
}
