% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSS_table.R
\name{BSS_table}
\alias{BSS_table}
\title{BSS_table}
\usage{
BSS_table(x, NBSS = F)
}
\arguments{
\item{x}{tsv file of a sample from EcoTaxa}

\item{NBSS}{Should it return NBSS or BSS ? False by default, return a BSS.}
}
\value{
Table of biovolume and abundance.
}
\description{
Return a table of biovolumes by size class and annotation for all types of biovolumes and abundance (elli, plain and riddled).
}
\examples{
BSS_table(x=tsv.file, NBSS=FALSE)
}
