% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bv_to_esd.R
\name{bv_to_esdum}
\alias{bv_to_esdum}
\title{bv_to_esdum}
\usage{
bv_to_esdum(x)
}
\arguments{
\item{x}{Biovolume in mm3}
}
\value{
ESD in micron
}
\description{
Compute ESD in micron from Biovolume in mm3
}
\examples{
bv_to_esdum(x=Biovolume)
}
