% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_obj-volume.R
\name{calc_ellps_vol}
\alias{calc_ellps_vol}
\title{Calculate the ellipsoid volume of a vignette}
\usage{
calc_ellps_vol(major, minor, pixel_mm)
}
\arguments{
\item{major}{the major axis, can be a vector}

\item{minor}{the minor axis, can be a vector}

\item{pixel_mm}{pixel to mm conversion}
}
\description{
Should give the raw input with pixels for the size values
}
