% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopart_par_conc.R
\name{calc_par_conc}
\alias{calc_par_conc}
\title{inner particle concentration calculator}
\usage{
calc_par_conc(par, min_esd, max_esd, pixel_mm, img_vol)
}
\arguments{
\item{par}{a par_df or ecopart_obj}

\item{min_esd}{minimum esd IN MM}

\item{max_esd}{maximum esd IN MM}

\item{pixel_mm}{pixel conversion}

\item{img_vol}{image volume size}
}
\description{
Get the particle volumetric concentraiton from a par file
}
