% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopart_par_conc.R
\name{calc_par_conc_bin}
\alias{calc_par_conc_bin}
\title{Inside calculation for particle counts by bin}
\usage{
calc_par_conc_bin(par, bin_limits, pixel_mm, img_vol)
}
\arguments{
\item{par}{a par df}

\item{bin_limits}{ESD limits for particle bin sizes in MM}

\item{pixel_mm}{pixel conversion}

\item{img_vol}{image volume size}
}
\description{
Calculate based on some bin size the particle concentration
}
