% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_obj-volume.R
\name{calc_sph_vol}
\alias{calc_sph_vol}
\title{Calculate Spherical Volume of vignette}
\usage{
calc_sph_vol(area, pixel_mm)
}
\arguments{
\item{area}{the area in pixels for a vignette}

\item{pixel_mm}{pixel to mm conversion}
}
\description{
Calculate Spherical Volume of vignette
}
