% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopart_zoo_conc.R
\name{cast_assign}
\alias{cast_assign}
\title{Inside function to assign casts for lapply}
\usage{
cast_assign(cast_name, ecopart_obj, ...)
}
\arguments{
\item{cast_name}{name of the cast to get concentration of}

\item{ecopart_obj}{an ecopart object list}

\item{...}{pass to \code{\link[=bin_taxa]{bin_taxa()}} for func and func_col}
}
\description{
Inside function to assign casts for lapply
}
