% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bin-helpers.R
\name{check_custom}
\alias{check_custom}
\title{check_custom - a function to check the custom range input}
\usage{
check_custom(custom_range, max_d)
}
\arguments{
\item{custom_range}{user defined range}

\item{max_d}{max depth from df}
}
\description{
check_custom - a function to check the custom range input
}
