% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_metadata.R
\name{check_metadata}
\alias{check_metadata}
\title{check_metadata}
\usage{
check_metadata(path, output)
}
\arguments{
\item{path}{path to the directory containing all the tsv files}

\item{output}{path to the output directory where the results are saved}
}
\value{
return a metadata resume of the directory and save the results
}
\description{
Check the metadata of all samples in a project.
}
