% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimanager.R
\name{chk_input}
\alias{chk_input}
\title{chk_input - helper to check input against known values}
\usage{
chk_input(input, known)
}
\arguments{
\item{input}{the user-listed name}

\item{known}{a vector of known names}
}
\description{
chk_input - helper to check input against known values
}
