% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_bv.R
\name{compute_bv}
\alias{compute_bv}
\title{compute_bv}
\usage{
compute_bv(path, output, metadata = NULL)
}
\arguments{
\item{path}{path to the directory containing all the tsv files}

\item{output}{path to the output directory where the results are saved}

\item{metadata}{metadata table generated by "check_metadata" or similar (if NULL, by default, use original metadata)}
}
\value{
Compute biovolumes for all the samples and save the new tsv files.
}
\description{
Compute all types of biovolumes (elli, plain, riddled) for each objects of one sample and save the new tsv files in the output directory.
Also compute conversion factor (by acquisition and by sample).
Convert the metadata to homogeneous units and replace NA by 1.
}
\examples{
compute_bv(path=path of the file, output=path where to save results, metadata=NULL)

}
