% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_size-spectra-tools.R
\name{count_size_classes}
\alias{count_size_classes}
\title{Count he number of observations in each size bin}
\usage{
count_size_classes(df)
}
\arguments{
\item{df}{a dataframe with a size_class column}
}
\description{
Count he number of observations in each size bin
}
