% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_size-spectra-tools.R
\name{create_log_size_classes}
\alias{create_log_size_classes}
\title{Generate size limits based on a log scale}
\usage{
create_log_size_classes(min_size, max_size, base = 2)
}
\arguments{
\item{min_size}{the minimum size class}

\item{max_size}{the maximum size class}

\item{base}{base of logarithm for bin width. Default 2}
}
\description{
Useful interior function for NBSS or particle size spectrum calcs
}
