% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_size-spectra-tools.R
\name{create_width_size_classes}
\alias{create_width_size_classes}
\title{Generate size limits based on a fixed width}
\usage{
create_width_size_classes(min_size, max_size, width)
}
\arguments{
\item{min_size}{the minimum size class}

\item{max_size}{the maximum size class}

\item{width}{the width of the approach -- needs to be selected}
}
\description{
Useful interior function for NBSS or particle size spectrum calcs
}
