% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoImport.R
\name{ecoImport}
\alias{ecoImport}
\title{ecoImport - quick way to read tsv files from ecotaxa}
\usage{
ecoImport(fileName, trim = F, customTrim = NULL)
}
\arguments{
\item{fileName}{path to the tsv export file from ecotaxa}

\item{trim}{whether or not to trim to certain columns only, default F}

\item{customTrim}{names of columns to keep in a trim}
}
\description{
ecoImport - quick way to read tsv files from ecotaxa
}
\examples{

}
