% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopart_example.R
\docType{data}
\name{ecopart_example}
\alias{ecopart_example}
\title{ecopart_example}
\format{
An object of class \code{'ecopart_obj'}; see \code{\link[EcotaxaTools]{ecopart_import}}.
}
\source{
\href{https://ecopart.obs-vlfr.fr/}{ecopart}
}
\usage{
data(ecopart_example)
}
\description{
Data were collected and imported using default features in ecopart_import
This dataset has 25 casts, although is missing a zoo_df for bats361_ctd18
}
\keyword{datasets}
