% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopart_import.R
\name{ecopart_file_check}
\alias{ecopart_file_check}
\title{Warning message for unmatched PAR and ZOO files}
\usage{
ecopart_file_check(par_fnames, zoo_fnames)
}
\arguments{
\item{par_fnames}{the par names}

\item{zoo_fnames}{the zoo names}
}
\description{
This is a warning message and provides indices to remove PAR or ZOO files
that don't have matches
}
