% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_init-functions.R
\name{ecopart_obj}
\alias{ecopart_obj}
\title{instatiate an ecopart_obj}
\usage{
ecopart_obj()
}
\description{
Sometimes you might need to instatiate new objects. If you just make a list,
it can be tedious to resurrect the class structure for the new object.
You can use this to instatiate an empty object
}
