% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_functions.R
\name{ellps_vol}
\alias{ellps_vol}
\title{ellps_vol - calculate volume of an ellipsoid shape}
\usage{
ellps_vol(df)
}
\arguments{
\item{df}{\itemize{
\item the data frame to feed it
}}
}
\description{
This function takes a data frame of ecotaxa output and returns a vector of ellipsoid shape
}
\details{
Must have column headers formatted. This assumes you are using the pixels input
it will return it in mm. If you converted to mm already - this doesn't work
unless you renamed converted columns
}
