% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimanager.R
\name{findMultiple}
\alias{findMultiple}
\title{findMultiple - locate all multiple cateogories}
\usage{
findMultiple(df)
}
\arguments{
\item{df}{a data frame with object_annotation_category}
}
\description{
Helper function to identify all types of multiple classes in
object_annotation_category
}
