% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_read-etx.R
\name{format_header}
\alias{format_header}
\title{format_header - a function to format headers}
\usage{
format_header(df)
}
\arguments{
\item{df}{a data frame to read the headers}
}
\description{
Takes the ecotaxa output and formats certain headers
}
