% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_etx.R
\name{format_headers}
\alias{format_headers}
\title{format_headers - a function to format headers}
\usage{
format_headers(df)
}
\arguments{
\item{df}{a data frame to read the headers}
}
\description{
Takes the ecotaxa output and formats certain headers
}
