% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopart_get_all.R
\name{get_all}
\alias{get_all}
\title{Get all for some column}
\usage{
get_all(zoo, col_name, pixel_conv = F)
}
\arguments{
\item{zoo}{an ecopart object or zoo_list}

\item{col_name}{a column name in the zoo_dfs}

\item{pixel_conv}{if using a size metric, should you convert the pixel size?}
}
\description{
Get all for some column
}
\examples{
# Get depth information
ecopart_example |> 
  get_all('depth_including_offset') |> 
  mean()

# Get size range
ecopart_example |> 
  get_all('esd', pixel_conv = TRUE) |> 
  range()
}
