% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_formatting.R
\name{get_bin_limtis}
\alias{get_bin_limtis}
\title{get_bin_limits
This function pulls bin limits from the categorys made by the cut() function}
\usage{
get_bin_limtis(input)
}
\arguments{
\item{input}{the vector of characters - should be a db from bin_taxa typically}
}
\description{
get_bin_limits
This function pulls bin limits from the categorys made by the cut() function
}
\author{
Alex Barth
}
