% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_col_name}
\alias{get_col_name}
\title{get_col_name}
\usage{
get_col_name(df, goal_name = NULL, possible = NULL)
}
\arguments{
\item{df}{the data frame, exported from ecopart or ecotaxa}

\item{goal_name}{the name of column}

\item{possible}{possible column names}
}
\description{
A function to choose which column to use
This is a hacky fix to the issue that column names switch
}
