% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopart_vol_bin.R
\name{get_ecopart_vol}
\alias{get_ecopart_vol}
\title{Get Volumes from Par files}
\usage{
get_ecopart_vol(ecopart_obj)
}
\arguments{
\item{ecopart_obj}{a list from ecopart_import()}
}
\description{
This takes in the par_files from an ecopart object
then returns the volume sampled in 1m depth bins
}
\seealso{
\code{\link[=uvp_par_conc]{uvp_par_conc()}}
}
\author{
Alex Barth
}
