% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_obj-volume.R
\name{get_zoo_df_features}
\alias{get_zoo_df_features}
\title{Get major and minor axis from zoo_df}
\usage{
get_zoo_df_features(zoo_df)
}
\arguments{
\item{zoo_df}{the zoo_df}
}
\description{
Get major and minor axis from zoo_df
}
