% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_style.R
\name{gg_cbb_col}
\alias{gg_cbb_col}
\title{Colorblind friendly pallete}
\usage{
gg_cbb_col(n)
}
\description{
Define a color-blind friendly palette for up to 8 items.
Colors are from the r-cookbook:
\ref{http://www.cookbook-r.com/Graphs/Colors_(ggplot2)/#a-colorblind-friendly-palette}
More than 8 will be defined by ggplot defaults.
}
\examples{
gg_cbb_col(3)

}
