% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_metadata.R
\name{graph.metadata}
\alias{graph.metadata}
\title{graph.metadata}
\usage{
graph.metadata(metadata)
}
\arguments{
\item{metadata}{metadata table generated by "check_metadata"}
}
\value{
A set of graphics to resume the project metadata.
}
\description{
A set of graphics to resume the project metadata.
Return a map of the samples, date and time, and metadata comparison between all samples.
}
