% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_project.R
\name{graph.project}
\alias{graph.project}
\title{graph.project}
\usage{
graph.project(x, metadata, taxo, bv.type = "elli", living.only = T)
}
\arguments{
\item{x}{bss data.table produced by "BSS_table"}

\item{metadata}{metadata table generated by "check_metadata"}

\item{taxo}{taxonomic table generated with "add_trophic" and "add_taxo"}

\item{bv.type}{elli, plain or riddled biovolume. default is "elli"}

\item{living.only}{TRUE by default, for NBSS plots}
}
\value{
A set of graphics to resume the project.
}
\description{
Print ggplot graphics to resume the project (comparison of all samples).
Return :
1. Text resume of all the project
2. live/not-live/temporary BV and AB graph
3. Not-Living and Living (based on the OTU database) BV and AB graph
4. NBSS (color and curve)
5. Relative NBSS
6. Shannon index
}
\examples{
graph.project(x=bss table of all the project, metadata, taxo=zoo.csv, bv.type="elli", living.only=TRUE)
}
