% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_sample.R
\name{graph.sample}
\alias{graph.sample}
\title{graph.sample}
\usage{
graph.sample(x, metadata, taxo, bv.type = "elli", living.only = T)
}
\arguments{
\item{x}{BSS table for one sample}

\item{metadata}{metadata table generated by "check_metadata"}

\item{taxo}{taxonomic table generated with "add_trophic" and "add_taxo"}

\item{bv.type}{elli, plain or riddled biovolume}

\item{living.only}{TRUE by default}
}
\value{
A set of graphics to resume the sample.
}
\description{
Return ggplot graphics to resume one sample.
1. Text resume
2. Relative AB and BV
3. Trophic pyramid
4. BSS and NBSS
5. Relative NBSS
6. Map
}
