% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_integration-tools.R
\name{integrate_all}
\alias{integrate_all}
\title{Integrate all taxa in a group}
\usage{
integrate_all(df, need_format = F, ...)
}
\arguments{
\item{df}{input data frame}

\item{need_format}{logical input to determine if needed to format}
}
\description{
Apply trapz integrate to a list of data frames
}
