% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_integration-tools.R
\name{integration_matcher}
\alias{integration_matcher}
\title{Match df to integrate}
\usage{
integration_matcher(df, ...)
}
\arguments{
\item{df}{the input data frame}

\item{...}{arguments to pass to integrate}
}
\description{
Interior function will match up a data frame for integration
}
