% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_integration-tools.R
\name{intg_to_tib}
\alias{intg_to_tib}
\title{Switch integration list to a dataframe}
\usage{
intg_to_tib(intg_list)
}
\arguments{
\item{intg_list}{the integration list}
}
\description{
This takes an etx_depth_integration_list and will return a dataframe
}
\author{
Alex Barth
}
