% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_class_checks.R
\name{is.etx_class}
\alias{is.etx_class}
\title{Check classes of objects}
\usage{
is.etx_class(input, classCheck)
}
\arguments{
\item{input}{the input of object to class check}

\item{classCheck}{the class to check against}
}
\description{
These are a suite of functions which are to check the class of objects
Mainly these are for allowing for functions to accept different types
}
