% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_integration-tools.R
\name{lin_interp}
\alias{lin_interp}
\title{Linear interpolation between sets of points}
\usage{
lin_interp(x, y, min_x = 0, max_x)
}
\arguments{
\item{x}{typically this should be depth}

\item{y}{typically this should be n/m3 or mg/m3}

\item{min_x}{expands the first y value to this value default at 0}

\item{max_x}{expands the last y value to this value}
}
\description{
Linear interpolation between sets of points
}
\author{
Alex Barth
}
