% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_list-to-tib.R
\name{list_to_tib}
\alias{list_to_tib}
\title{Make a tibble from a list - with names as a new column}
\usage{
list_to_tib(list, new_col_name = "group")
}
\arguments{
\item{list}{the input list - should have all elements be a df}

\item{new_col_name}{the column name to correspond to list names}
}
\description{
This function just calls rbind on a whole list but will first
make a new column in the dataframe, corresponding to the names of the list
}
\author{
Alex Barth
}
