% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopart_add_zoo.R
\name{mod_zoo}
\alias{mod_zoo}
\title{Modify a list of ecotaxa tsv dfs}
\usage{
mod_zoo(zoo_list, func, ...)
}
\arguments{
\item{zoo_list}{a list of ecotaxa tsvs or an ecopart-object}

\item{func}{the modifier function to apply}

\item{...}{additional arguments to pass to function}
}
\description{
Apply a modifying function to a list of dataframes or ecopart object
}
\author{
Alex Barth
}
