% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_multimanager.R
\name{multi_manager}
\alias{multi_manager}
\title{multi_manager - a tool to re-label vignettes with multiple individuals}
\usage{
multi_manager(path = NULL, morpho_include = T)
}
\arguments{
\item{path}{location of file to choose, if not entered, a box will open to select file}

\item{morpho_include}{if true, will create a column of boolean values. True indicates that row will be included in morpho-measurements. New rows from multimanager default to F.}
}
\description{
This function is a tool which facilitates quick renaming of vignettes
with multiple organisms
It will select all multiple-object vignettes then request user input for the
names of the new file
For each ID entered, it will create a new line with that estimate
Note that there will
}
