% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimanager.R
\name{multiManager}
\alias{multiManager}
\title{multiManager - a tool to re-label vignettes with multiple individuals}
\usage{
multiManager(path, morphoInclude = T, openImage = T, imagePath = NULL)
}
\arguments{
\item{path}{location of file to choose, if not entered, a box will open to select file}

\item{morphoInclude}{if true, will create a column of boolean values. True indicates that row will be included in morpho-measurements. New rows from multimanager default to F.}

\item{openImage}{if true, will find and open images. Requires file connection}

\item{imagePath}{if open image is true, a path to the image folder must be provided}

\item{removeMultiple}{if true, this will delete the multiple< lines}
}
\description{
This function is a tool which facilitates quick renaming of vignettes
with multiple organisms
It will select all multiple-object vignettes then request user input for the
names of the new file
For each ID entered, it will create a new line with that estimate
Note that there will
}
