% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_rename.R
\name{name_match}
\alias{name_match}
\title{Find a match to list}
\usage{
name_match(taxo_hier, poss_match)
}
\arguments{
\item{taxo_hier}{taxonomoy hierarchy}

\item{poss_match}{list of possible matches}
}
\description{
Inside piece to broad_names
}
