% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_rename.R
\name{names_drop}
\alias{names_drop}
\title{Drop certain names from the ecotaxa dataframe}
\usage{
names_drop(df, drop_names, drop_children = F)
}
\arguments{
\item{df}{an ecotaxa-tsv style data frame}

\item{drop_names}{character vector of which names to drop}

\item{drop_children}{option to drop taxonomic children}
}
\description{
This is a modifier which will remove rows which contain matching function values
}
\examples{
# Droping all not-living categories:
ecopart_example <- ecopart_example |> mod_zoo(func = names_drop,
                                              drop_names = 'not-living',
                                              drop_children = TRUE)

# Can be done with a single DF:
# Drop all artefacts
cast1_no_arte <- ecopart_example$zoo_files$bats361_ctd1 |> 
  names_drop('artefact',drop_children = TRUE)
}
\author{
Alex Barth
}
