% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_rename.R
\name{names_keep}
\alias{names_keep}
\title{Keep only names from the ecotaxa dataframe}
\usage{
names_keep(df, keep_names)
}
\arguments{
\item{df}{an ecotaxa-tsv style data frame}

\item{keep_names}{character vector of which names to drop}
}
\description{
This is a modifier which will remove rows that don't contain values provided
An alternavite is names_drop
}
\author{
Alex Barth
}
