% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_rename.R
\name{names_to}
\alias{names_to}
\title{Assign names to a possible list}
\usage{
names_to(df, new_names, suppress_print = F)
}
\arguments{
\item{df}{an ecotaxa tsv-style dataframe}

\item{new_names}{a vector of character names to move towards default}

\item{suppress_print}{set to true to stop messages of matches}
}
\description{
Assign names to a possible list
}
\author{
Alex Barth
}
