% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBSS_plot.R
\name{NBSS.plot}
\alias{NBSS.plot}
\title{NBSS.plot}
\usage{
NBSS.plot(x, taxon, bv.type, samples)
}
\arguments{
\item{x}{bss data.table produced by "BSS_table"}

\item{taxon}{taxon you want to plot. Choose "total" if you want to see the global NBSS.}

\item{bv.type}{elli, plain or riddled biovolume. default is "elli"}

\item{samples}{samples you want to plot.}
}
\value{
A NBSS ggplot
}
\description{
NBSS plot function.
}
