% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{nearest}
\alias{nearest}
\title{nearest}
\usage{
nearest(number, vector)
}
\arguments{
\item{number}{the number to check}

\item{vector}{a numeric vector to match values to}
}
\description{
A function to assign a numeric bin between a range
}
\examples{
nearest(74.2,seq(0,100,5))

}
