% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_size-spectra-tools.R
\name{numeric_spectra}
\alias{numeric_spectra}
\title{Calculate the numeric particle spectra}
\usage{
numeric_spectra(df, vol_L, needs_format = FALSE)
}
\arguments{
\item{df}{a size.count.df or data frame with size_class column}

\item{vol_L}{the volumn in L, will be converted to cm3}

\item{needs_format}{if TRUE, will bin_format}
}
\description{
Calculate numeric particle spectra
}
