% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binByTools.R
\name{order_bins}
\alias{order_bins}
\title{order_bins - a function to order bins correctly}
\usage{
order_bins(rdf)
}
\arguments{
\item{rdf}{\itemize{
\item the rdf right before return
}}
}
\description{
order_bins - a function to order bins correctly
}
