% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopart_par_spectra.R
\name{par_df_spectra}
\alias{par_df_spectra}
\title{Calculate numeric spectra in depth bins of par_df}
\usage{
par_df_spectra(par_df, depth_breaks, pixel_mm, img_vol, ...)
}
\arguments{
\item{par_df}{the original par_df}

\item{depth_breaks}{the depth breaks}

\item{pixel_mm}{pixel conversion}

\item{img_vol}{image to volume}

\item{...}{arguments to pass to \code{assign_spectra_bins}}
}
\description{
THis is an interior function for a single iteration
}
