% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopart_plot_par.R
\name{particle_plot}
\alias{particle_plot}
\title{Plot a particle diagram}
\usage{
particle_plot(
  par,
  min_esd = 0,
  max_esd = 100,
  open_windows = TRUE,
  bin_limits = NULL,
  pixel_mm = NULL,
  img_vol = NULL
)
}
\arguments{
\item{par}{a par_df or ecopart_obj}

\item{min_esd}{minimum esd IN MM}

\item{max_esd}{maximum esd IN MM}

\item{open_windows}{set to true will open plots for each profile}

\item{bin_limits}{optional for returning size-binned particle concentration}

\item{pixel_mm}{pixel conversion}

\item{img_vol}{image volume size}
}
\description{
Plot a particle diagram
}
