% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkscoperoutine-package.R
\docType{package}
\name{pkscoperoutine-package}
\alias{pkscoperoutine}
\alias{pkscoperoutine-package}
\title{pkscoperoutine: EcoTaxa tools for PlanktoScope}
\description{
This package can compute biovolumes from a ecotaxa tsv file and gives an overview of the data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pkostyrka/pkscoperoutine}
  \item Report bugs at \url{https://github.com/pkostyrka/pkscoperoutine/issues}
}

}
\keyword{internal}
