% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopart_merge_casts.R
\name{pool_casts}
\alias{pool_casts}
\title{Inside cast pooling}
\usage{
pool_casts(cast_names, cast_list)
}
\arguments{
\item{cast_names}{the names of the casts which should be pooled}

\item{cast_list}{the list of all cast dataframe}
}
\description{
This function will take either zoo files and/or par files and put them into
Similar bins based on their names
}
\author{
Alex Barth
}
