% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_uvp-tools.R
\name{raw_bru_import}
\alias{raw_bru_import}
\title{Read in raw bru files}
\usage{
raw_bru_import(proj_path, cast_names = NULL)
}
\arguments{
\item{proj_path}{the path to the project root folder}

\item{cast_names}{names of casts (in work folder). If null will read all}
}
\description{
Read in raw bru files
}
