% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_uvp-tools.R
\name{raw_dat_import}
\alias{raw_dat_import}
\title{Read in raw dat files}
\usage{
raw_dat_import(proj_path, cast_names = NULL, trim_sensor = T)
}
\arguments{
\item{proj_path}{the path to the project root folder}

\item{cast_names}{names of casts (in work folder). If null will read all}

\item{trim_sensor}{remove excess sensor/summary data}
}
\description{
Sometimes it is useful to read in the dat files from
a UVP project. Note, that this does not include the functionality
to read directly from the raw folder. It is not advisable to read
directly from the raw file. However, if you want to evaluate the firstImage
detection, then you'll need to do that.
}
\details{
Also be aware the depth does not include offset
}
