% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_uvp-tools.R
\name{raw_merge_brudat}
\alias{raw_merge_brudat}
\title{Merge raw_dat and raw_bru files}
\usage{
raw_merge_brudat(bru, dat)
}
\arguments{
\item{bru}{bru files}

\item{dat}{dat files}
}
\description{
Merge raw_dat and raw_bru files
}
