% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_read-etx.R
\name{read_etx}
\alias{read_etx}
\title{read_etx - a function to read tsv from ecotaxa}
\usage{
read_etx(path, format_headers = T)
}
\arguments{
\item{path}{the file to open}

\item{format_headers}{\itemize{
\item if true, it will remove object_. This is required for code compatibiility
}}
}
\description{
Honestly, tired of making sure to use ecotaxar package so I copied this one
}
