% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_uvp-tools.R
\name{read_raw_bru}
\alias{read_raw_bru}
\title{Interior bru reader}
\usage{
read_raw_bru(cast_name, proj_path)
}
\arguments{
\item{cast_name}{the cast to read}

\item{proj_path}{the path to the project root folder}
}
\description{
Interior bru reader
}
