% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_uvp-tools.R
\name{read_raw_datfile}
\alias{read_raw_datfile}
\title{Read raw dat}
\usage{
read_raw_datfile(cast_name, proj_path, trim_sensor = TRUE)
}
\arguments{
\item{cast_name}{the cast to read}

\item{proj_path}{the path to the project root folder}

\item{trim_sensor}{trim all sensor data}
}
\description{
Interior for single raw dat file
}
