% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_rel-taxa.R
\name{rel_abundance}
\alias{rel_abundance}
\title{Get the percent contribution of taxa}
\usage{
rel_abundance(df)
}
\arguments{
\item{df}{an ecotaxa .tsv style dataframe}
}
\description{
This function is for the inner bits of rel_taxa
}
\author{
Alex Barth
}
