% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_rel-taxa.R
\name{rel_taxa}
\alias{rel_taxa}
\title{Get the percent contribution of taxa}
\usage{
rel_taxa(ecodf)
}
\arguments{
\item{ecodf}{a list, data frame, or ecopart_obj}
}
\description{
This will take either a list or a df and get the relative abundance
}
\examples{
# For a single cast:
ecopart_example$zoo_files$bats361_ctd1 |> rel_taxa()

# For multiple casts:
rel_taxa(ecopart_example)
rel_taxa(ecopart_example$zoo_files)
}
\author{
Alex Barth
}
