% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_routine.R
\name{routine_planktoscope_ecotaxa}
\alias{routine_planktoscope_ecotaxa}
\title{routine_planktoscope_ecotaxa}
\usage{
routine_planktoscope_ecotaxa()
}
\value{
Biovolumes, dataframes and resume tables, figures to check the data.
}
\description{
A routine to compute biovolumes and save updated csv files. Save also table resume for each taxonomic group and figures to check the data.
}
