% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopart_plot_par.R
\name{single_par_plot}
\alias{single_par_plot}
\title{Plot a single particle diagram -- total concentration}
\usage{
single_par_plot(par_conc_df, ...)
}
\arguments{
\item{par_conc_df}{a particle conc df}

\item{...}{arguments to pass to \code{geom_smooth} can set gam or loess, span, etc}
}
\description{
Plot a single particle diagram -- total concentration
}
