% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_functions.R
\name{sph_vol}
\alias{sph_vol}
\title{sph_vol - calculate volume of a spherical object}
\usage{
sph_vol(df)
}
\arguments{
\item{df}{\itemize{
\item an ecotaxa dataframe
}}
}
\description{
This function calculates the volume of an object if it were a sphere
Input an ecotaxa df and get back a vector of spherical volums
}
