% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopart_split_obj.R
\name{split_ecopart_obj}
\alias{split_ecopart_obj}
\title{Split an ecopart object based on some meta_data}
\usage{
split_ecopart_obj(ecopart_obj, split_by)
}
\arguments{
\item{ecopart_obj}{ecopart object to split}

\item{split_by}{character corresponding to a metadata column}
}
\description{
Split an ecopart object based on some meta_data
}
