% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvp_conc.R
\name{sum_casts}
\alias{sum_casts}
\title{Tool to add similar uvp_casts based on some bin}
\usage{
sum_casts(cast_names, zoo_list, vol_list, custom_range, ...)
}
\arguments{
\item{cast_names}{character vector of cast names in the list values}

\item{zoo_list}{an ecopart zooplankton list}

\item{vol_list}{an ecopart volume list}

\item{custom_range}{custom depth bin limits}

\item{...}{extra arguments to pass to \code{bin_taxa()}}
}
\description{
See also \code{sum_casts_conc()}
}
\author{
Alex barth
}
