% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvp_conc.R
\name{sum_casts_conc}
\alias{sum_casts_conc}
\title{Calculation of concentraiton from summed casts}
\usage{
sum_casts_conc(cast_names, zoo_list, vol_list, custom_range, ...)
}
\arguments{
\item{cast_names}{character vector of cast names in the list values}

\item{zoo_list}{an ecopart zooplankton list}

\item{vol_list}{an ecopart volume list}

\item{custom_range}{custom depth bin limits}

\item{...}{other arguments to pass to \code{bin_taxa()}}
}
\description{
see also \code{sum-casts}
}
\author{
Alex Barth
}
