% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_integration-tools.R
\name{trapz_integarate}
\alias{trapz_integarate}
\title{Trapezoidal integration}
\usage{
trapz_integarate(x, y, min_x, max_x, ...)
}
\arguments{
\item{x}{typically this should be depth}

\item{y}{typically this should be n/m3 or mg/m3}

\item{min_x}{expands the first y value to this value}

\item{max_x}{expands the last y value to this value}

\item{...}{arguements to pass to \code{integrate} : most helpful to increase subdivisons or rel.tol}
}
\description{
This exists in a much more comprehensive way in dan kelley's oce
but i hate using too many packages
}
\author{
Alex Barth
}
