% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopart_trim-obj.R
\name{trim_ecopart_depth}
\alias{trim_ecopart_depth}
\title{Trim an ecopart object by depth}
\usage{
trim_ecopart_depth(ecopart_obj, depth, operator = "less")
}
\arguments{
\item{ecopart_obj}{an ecopart_obj}

\item{depth}{the depth limit}

\item{operator}{'less','greater','lessequal', or 'greaterequal'}
}
\description{
This is a quick function to filter an ecopart object by depth
it runs on both trim_zoo and trim_par
}
