% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopart_trim-obj.R
\name{trim_par_inner}
\alias{trim_par_inner}
\title{Inside par trimmer}
\usage{
trim_par_inner(par_list, trim_cat, trim_amt, operator)
}
\arguments{
\item{par_list}{either a par list}

\item{trim_cat}{the category to trim by}

\item{trim_amt}{the amount to trim by}

\item{operator}{either 'less','greater','lessequal', or 'greaterequal'}
}
\description{
Inside par trimmer
}
