% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopart_trim-obj.R
\name{trim_zoo}
\alias{trim_zoo}
\title{Trim zoo_list}
\usage{
trim_zoo(zoo_list, trim_cat, trim_amt, operator)
}
\arguments{
\item{zoo_list}{a zoo_list}

\item{trim_cat}{the category to trim by}

\item{trim_amt}{the amount to trim by}

\item{operator}{either 'less','greater','lessequal', or 'greaterequal'}
}
\description{
A simple function to trim zoo_list. see also trim_par
}
