% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopart_trim-obj.R
\name{trim_zoo_inner}
\alias{trim_zoo_inner}
\title{Inside zoo_list trimmer
A simple function to trim par list based on some criteria
syntax is similar to lapply and users must define a trim function}
\usage{
trim_zoo_inner(zoo_list, trim_cat, trim_amt, operator)
}
\arguments{
\item{zoo_list}{a zoo_list}

\item{trim_cat}{the category to trim by}

\item{trim_amt}{the amount to trim by}

\item{operator}{either 'less','greater','lessequal', or 'greaterequal'}
}
\description{
Inside zoo_list trimmer
A simple function to trim par list based on some criteria
syntax is similar to lapply and users must define a trim function
}
