% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archived_tsvTrim.R
\name{tsv_trim}
\alias{tsv_trim}
\title{tsv_trim - an automated trimming tool for ecotaxa tsv's}
\usage{
tsv_trim(df, trim_mode = "Minimal", custom_vector = NULL)
}
\arguments{
\item{df}{the tsv file from ecoImport}

\item{trim_mode}{set to "Minimal", "Morpho","sample_info", "Custom"}

\item{custom_vector}{if "Custom" trim_mode provide a vector of column names}
}
\description{
This function has a built-in list of columns to trim tsv files from ecotaxa
I retired my ecoImport funtion after discovering ecotaxar's much better
read_ecotaxa_tsv funcitons
}
