% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsvTrim.R
\name{tsvTrim}
\alias{tsvTrim}
\title{tsvTrim - an automated trimming tool for ecotaxa tsv's}
\usage{
tsvTrim(df, trimMode = "Minimal", customVector = NULL)
}
\arguments{
\item{df}{the tsv file from ecoImport}

\item{trimMode}{Which preset trim set "Minimal", "Morpho", or "Custom"}

\item{customVector}{if "Custom" trimMode provide a vector of column names}
}
\description{
This function has a built-in list of columns to trim tsv files from ecotaxa
I retired my ecoImport funtion after discovering ecotaxar's much better
read_ecotaxa_tsv funcitons
}
