% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvp_conc.R
\name{uvp_conc}
\alias{uvp_conc}
\title{Calculate UVP Concentration}
\usage{
uvp_conc(cast_names, zoo_list, vol_list, custom_range, ...)
}
\arguments{
\item{cast_names}{a character vector of cast names to loop over}

\item{zoo_list}{list of ecopart zooplankton data}

\item{vol_list}{list of volume ecopart data}

\item{custom_range}{range of vectors to break into bins}

\item{...}{arguments to pass to \code{bin_taxa}}
}
\description{
This script takes the UVP data and calculates densities
it requries output from raw ecopart export for both taxa and volume
These files are automatically imported by the ecopart_import
}
\details{
Eventually I will need to update this function since i don't want looping inside functions
}
\author{
Alex Barth
}
