% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopart_par_conc.R
\name{uvp_par_conc}
\alias{uvp_par_conc}
\title{Calculate particle concentration}
\usage{
uvp_par_conc(
  par,
  min_esd = 0,
  max_esd = 100,
  pixel_mm = NULL,
  img_vol = NULL,
  bin_limits = NULL
)
}
\arguments{
\item{par}{a par_df or ecopart_obj}

\item{min_esd}{minimum esd IN MM}

\item{max_esd}{maximum esd IN MM}

\item{pixel_mm}{pixel conversion}

\item{img_vol}{image volume size}

\item{bin_limits}{optional for returning size-binned particle concentration}
}
\description{
caculate the particle concentration for a given size class
The bulk of this code lives in \code{calc_par_conc} and this is a wrapper
}
