% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopart_par_spectra.R
\name{uvp_par_spectra}
\alias{uvp_par_spectra}
\title{Number Size Spectra by bins}
\usage{
uvp_par_spectra(
  ecopart_obj,
  depth_breaks,
  pixel_mm = NULL,
  img_vol = NULL,
  ...
)
}
\arguments{
\item{ecopart_obj}{the ecopart_obj}

\item{depth_breaks}{the depth breaks}

\item{pixel_mm}{pixel conversion}

\item{img_vol}{image to volume}

\item{...}{arguments to pass to \code{assign_spectra_bins}}

\item{method}{the method to by by}
}
\description{
Cacluate the numerical size spectrum by depth bins
}
