% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopart_par_spectra.R
\name{uvp_split_par}
\alias{uvp_split_par}
\title{Split par file based on depth bins}
\usage{
uvp_split_par(par_df, depth_breaks)
}
\arguments{
\item{par_df}{the particle file}

\item{depth_breaks}{a vector of depth breaks}
}
\description{
Split par file based on depth bins
}
