% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopart_zoo_conc.R
\name{uvp_zoo_conc}
\alias{uvp_zoo_conc}
\title{Calculate UVP Zooplankton Concentration from an Ecopart Object}
\usage{
uvp_zoo_conc(ecopart_obj, cast_name = NULL, breaks, ...)
}
\arguments{
\item{ecopart_obj}{an ecopart object list}

\item{cast_name}{the name (or names of a cast)}

\item{breaks}{a vector to break on}

\item{...}{pass into \code{\link[=bin_taxa]{bin_taxa()}}}
}
\description{
This function takes in an ecopart object and will calculate the
concentration for zooplankton data. This is a big wrapper function to
build and add functionality to \code{\link[=uvp_conc]{uvp_conc()}}.
}
\examples{
#concentration of a whole projects
zoo_conc <- ecopart_example |> uvp_zoo_conc(breaks = seq(0,1200,100))

# concentration for particular  taxa
zoo_conc <- ecopart_example |>
  add_zoo(func = names_to, col_name = 'name', 
          new_names = c('Chaetognatha','Copepoda',
                        'Eumalacostraca','living','not-living'),
          suppress_print = TRUE) |> # rename for simplicity
  mod_zoo(names_drop, drop_names = 'not-living') |> #remove detritus & artifact
  uvp_zoo_conc(cast_name = c('bats361_ctd1', 'bats361_ctd2'), 
               breaks = seq(0,1200,100)) |> # get concentration
  lapply(bin_format) #formatting

#take a peek:
head(zoo_conc$bats361_ctd1,8)
head(zoo_conc$bats361_ctd2,8)


#volumetric concentraiton for a single cast
cast1_copepod_conc <- ecopart_example |>
  mod_zoo(names_keep, keep_names = 'Copepoda', keep_children = TRUE) |>
  add_zoo(func = names_to, col_name = 'name', new_names = 'Copepoda', 
          suppress_print = TRUE) |>
  add_zoo(func = biovolume, col_name = 'biovol', shape = 'ellipsoid', 
          pixel_mm = unique(ecopart_example$meta$acq_pixel)) |>
  uvp_zoo_conc(cast_name = 'bats361_ctd12', breaks = seq(0,500,100)) |>
  bin_format()

cast1_copepod_conc
}
