% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archived_tsvTrim.R
\name{vector_select}
\alias{vector_select}
\title{vector_select - background mechanism for storing trim factors}
\usage{
vector_select(trim_mode)
}
\arguments{
\item{trim_mode}{set to "Minimal" or "Morpho" unless local selections added}
}
\description{
vector_select - background mechanism for storing trim factors
}
