% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsvTrim.R
\name{vectorSelect}
\alias{vectorSelect}
\title{vectorSelect - background mechanism for storing trim factors}
\usage{
vectorSelect(trimMode)
}
\arguments{
\item{trimMode}{set to "Minimal" or "Morpho" unless local selections added}
}
\description{
vectorSelect - background mechanism for storing trim factors
}
