% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_obj-volume.R
\name{VignetteVolume}
\alias{VignetteVolume}
\alias{ellps_vol}
\alias{sph_vol}
\title{Calculate the volume of a vignette}
\usage{
ellps_vol(major, minor, pixel_mm)

sph_vol(major, minor, pixel_mm)
}
\arguments{
\item{major}{the major axis, can be a vector}

\item{minor}{the minor axis, can be a vector}

\item{pixel_mm}{pixel to mm conversion}
}
\description{
Should give the raw input with pixels for the size values
}
