% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol_sampled.R
\name{vol_sampled}
\alias{vol_sampled}
\title{vol_sampled - get the volume sampled from a UVP cast}
\usage{
vol_sampled(
  cast_name,
  path,
  method,
  custom_range,
  equal_step = 10,
  equal_number = 100
)
}
\arguments{
\item{cast_name}{\itemize{
\item the name of the cast
}}

\item{path}{\itemize{
\item where the dat and meta files are located
}}

\item{method}{"Equal Space", "Custom", "Equal Number" "Zooscan"}

\item{custom_range}{vector of ranges}

\item{equal_step}{create equal steps of this size}

\item{equal_number}{create this number of ranges}
}
\description{
vol_sampled - get the volume sampled from a UVP cast
}
