% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zooregroup_zooscan}
\alias{zooregroup_zooscan}
\title{zooregroup_zooscan}
\format{
## `zooregroup_zooscan`
}
\usage{
zooregroup_zooscan
}
\description{
A database with trophic levels and OTU.
#' object_annotation_hierarchy2 = object annotation hierarchy from ecotaxa with corrected delim >.
Type = OTU main type "Living, non-living, etc."
Sub_Type = OTU sub type "Diatoms, Copepods, etc."
Value = Trophic level "Predators = 3 ... Autotroph = 1, NA = 3.5 and none = -1"
}
\keyword{datasets}
