% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_taxo.R
\name{add.taxo}
\alias{add.taxo}
\title{add.taxo}
\usage{
add.taxo(object_hierarchy_list)
}
\arguments{
\item{object_hierarchy_list}{List of object annotation hierarchy.}
}
\value{
Return a taxonomic table.
}
\description{
Return a taxonomic table.
}
