% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_zoo.R
\name{add.zoo}
\alias{add.zoo}
\title{add.zoo}
\usage{
add.zoo(taxo)
}
\arguments{
\item{taxo}{OTU table generated with "add_taxo" containing trophic levels and main OTU groups.}
}
\value{
A taxonomic table with also trophic levels.
}
\description{
Add trophic levels and OTU from a database constructed within the package, or external.
Type = OTU main
Sub_Type = OTU sub type
Value = Trophic level
}
