% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopart_add-mod_zoo.R
\name{add_zoo}
\alias{add_zoo}
\title{Add column to all zooplankton files}
\usage{
add_zoo(zoo_list, func, col_name, ...)
}
\arguments{
\item{zoo_list}{an ecopart object list or list of ecotaxa-tsv (zoo_files)}

\item{func}{function to apply (must take df and return vector) see ellps_vol}

\item{col_name}{name to assign to new column}

\item{...}{to pass to \code{lapply()} if needed}
}
\description{
This function will add a column to all zooplankton files in either an ecopart-object
or just a list of ecotaxa_tsvs
}
\author{
Alex Barth
}
