% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_helper-functions.R
\name{any_in}
\alias{any_in}
\title{Any In}
\usage{
any_in(checkVect, refVect)
}
\arguments{
\item{checkVect}{the input vector}

\item{refVect}{the reference vector}
}
\description{
Simply a wrapper for \%in\% to check if there are any
}
