% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binByTools.R
\name{bin_by_df}
\alias{bin_by_df}
\title{bin_by_df - returns a data frame for counts by bin}
\usage{
bin_by_df(
  df,
  method,
  custom_range,
  equal_step = 10,
  equal_number = 100,
  secondary = NULL,
  secondary_breaks = NULL,
  return_list = F
)
}
\arguments{
\item{df}{the vector to include to be cut. Should be an ecotaxa tsv export}

\item{method}{"Equal Space", "Custom", "Equal Number" "Zooscan"}

\item{custom_range}{vector of ranges}

\item{equal_step}{create equal steps of this size}

\item{equal_number}{create this number of ranges}

\item{secondary}{a secondary characteristic (like esd) to split these on}

\item{secondary_breaks}{the range at which to break the secondary var}

\item{return_list}{a T/F to get back a list of secondary breaks}
}
\description{
will count number of observations for each bin then return a count dataframe
}
