% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bin-helpers.R
\name{bin_format}
\alias{bin_format}
\title{Format Bins}
\usage{
bin_format(df, col = "db")
}
\arguments{
\item{df}{a dataframe with a column formatted as bins}

\item{col}{the column formatted as bins, default db}
}
\description{
Takes a concentration object and will format the bins
to have min_d, max_d, and mp (mid_point)
}
\author{
Alex Barth
}
