% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bin-taxa.R
\name{bin_taxa}
\alias{bin_taxa}
\title{bin_taxa() - a modern rendition of bin_by_df
This version allows for summing over some factor like dry mass or volume}
\usage{
bin_taxa(df, depth_breaks, zooscan = F, func_col = "taxo_name", func = length)
}
\arguments{
\item{df}{a data frame should have a taxo_name, name, or object_annotation_category column}

\item{depth_breaks}{the limits of what bins should be made}

\item{zooscan}{if the df is a zooscan export set to true, default false}

\item{func_col}{a single character vector of the value to apply function to; esd, biomass, drymass, etc}

\item{func}{the function to apply}
}
\description{
It will sum for each individual in a depth bin by taxa
}
\author{
Alex Barth
}
