% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binByTools.R
\name{bin_taxa}
\alias{bin_taxa}
\title{bin_taxa() - a modern rendition of bin_by_df
This version allows for summing over some factor like dry mass or volume}
\usage{
bin_taxa(df, custom_range, zooscan = F, apply_func = F, func_col, func)
}
\arguments{
\item{df}{a data frame should have a taxo_name, name, or object_annotation_category column}

\item{custom_range}{the limits of what bins should be made}

\item{zooscan}{if the df is a zooscan export set to true, default false}

\item{apply_func}{bin_taxa will apply function to a specified column if set to true}

\item{func_col}{a single character vector of the value to apply function to; esd, biomass, drymass, etc}

\item{func}{the function to apply}
}
\description{
It will sum for each individual in a depth bin by taxa
}
\author{
Alex Barth
}
