% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_obj-volume.R
\name{biovolume}
\alias{biovolume}
\title{Get vignette biovolume}
\usage{
biovolume(input, shape, pixel_mm)
}
\arguments{
\item{input}{a zoo_df. use add_zoo for ecopart_objs}

\item{shape}{assume a sphere or ellipsoid}

\item{pixel_mm}{option to user specify pixel size in mm}
}
\description{
This function takes a df and will direct it
}
