% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_bv.R
\name{compute_bv}
\alias{compute_bv}
\title{compute_bv}
\usage{
compute_bv(path, output, metadata)
}
\arguments{
\item{path}{path to the directory containing all the tsv files}

\item{output}{path to the output directory where the results are saved}

\item{metadata}{metadata table generated by "check_metadata"}
}
\value{
Compute biovolumes for all the samples and save the new tsv files.
}
\description{
Compute biovolumes for all the samples and save the new tsv files.
}
