% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopart_import.R
\name{ecopart_import}
\alias{ecopart_import}
\title{reads in files from a folder of ecopart raw export}
\usage{
ecopart_import(dat_path)
}
\arguments{
\item{dat_path}{the path in your compute to the file}
}
\description{
This function requires a path to a folder with the raw output from ecopart
and is primarily designed for UVP export objects. It will return a list of lists
each list holds tibbles for each cast in a project for: particle, plankton, volume, meta
}
\author{
Alex Barth
}
