% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopart_vol_bin.R
\name{ecopart_vol_bin}
\alias{ecopart_vol_bin}
\title{Ecopart Volume Binning}
\usage{
ecopart_vol_bin(df, method, depth_breaks)
}
\arguments{
\item{df}{a data frame to enter}

\item{method}{"Equal Space", "Custom", "Equal Number"}

\item{depth_breaks}{limits for depth bins}
}
\description{
This function calculates the volume for each 1-m depth bin
}
