% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecotaxatools-package.R
\docType{package}
\name{ecotaxatools-package}
\alias{ecotaxatools}
\alias{ecotaxatools-package}
\title{ecotaxatools: EcoTaxa tools}
\description{
This package can compute biovolumes from a ecotaxa tsv file and gives an overview of the data.
}
\keyword{internal}
