% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimanager.R
\name{getSelection}
\alias{getSelection}
\title{getMultiIndex - helper function for user input}
\usage{
getSelection(df)
}
\description{
This calls for user input then asks which multiples to include
then it provides the index for all multiples of interest
}
