% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_project.R
\name{graph.project}
\alias{graph.project}
\title{graph.project}
\usage{
graph.project(x, metadata, taxo, bv.type = "elli", living.only = T)
}
\arguments{
\item{x}{bss data.table produced by "BSS_table"}

\item{metadata}{metadata table generated by "check_metadata"}

\item{taxo}{taxonomic table generated with "add_trophic" and "add_taxo"}

\item{bv.type}{elli, plain or riddled biovolume. default is "elli"}

\item{living.only}{TRUE by default}
}
\value{
A set of graphics to resume the project.
}
\description{
Return ggplot graphics to resume the project.
}
