% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_rename.R
\name{names_drop}
\alias{names_drop}
\title{Drop certain names from the ecotaxa dataframe}
\usage{
names_drop(df, drop_names)
}
\arguments{
\item{df}{an ecotaxa-tsv style data frame}

\item{drop_names}{character vector of which names to drop}
}
\description{
This is a modifier which will remove rows which contain matching function values
}
\author{
Alex Barth
}
