% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_rename.R
\name{names_keep}
\alias{names_keep}
\title{Keep only names from the ecotaxa dataframe}
\usage{
names_keep(df, keep_names, keep_children = F)
}
\arguments{
\item{df}{an ecotaxa-tsv style data frame}

\item{keep_names}{character vector of which names to drop}

\item{keep_children}{option to keep children of drop names}
}
\description{
This is a modifier which will remove rows that don't contain values provided
An alternavite is names_drop
}
\examples{
# keep only copepods:
copes_only <- ecopart_example |> mod_zoo(names_keep, keep_names = 'Copepoda',
                                         keep_children = T)

# Keep multiple labels:
rhiz <- ecopart_example |> mod_zoo(names_keep, 
                                   keep_names = c('Aulosphaeridae', 
                                                  'Aulacanthidae'),
                                   keep_children = T)

# with a single zoo_df
cope_cast1_df <- ecopart_example$zoo_files$bats361_ctd1 |> 
  names_keep(keep_names = 'Copepoda',keep_children = T)
}
\author{
Alex Barth
}
