% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_rename.R
\name{names_to}
\alias{names_to}
\title{Assign names based on taxo hierarchy}
\usage{
names_to(df, new_names, suppress_print = F)
}
\arguments{
\item{df}{an ecotaxa tsv-style dataframe}

\item{new_names}{a vector of character names to move towards default}

\item{suppress_print}{set to true to stop messages of matches}
}
\description{
This allows you to use the taxo hierarch to move up to a more broad name
}
\examples{
# Re-classifying to a broader group:
Copepod_obj <- ecopart_example |> 
  add_zoo(func = names_to, col_name = 'name',
          new_names = c('living', 'not-living', 'Copepoda'),
          suppress_print  = TRUE)

#working with a single zoo_df will return name vector
cast1_copes <- ecopart_example$zoo_files$bats361_ctd1 |> 
  names_to(new_names = c('living', 'not-living', 'Copepoda'))
# how many copepods in cast 1?
cast1_copes |> table()
}
\author{
Alex Barth
}
