% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_routine.R
\name{routine_planktoscope_ecotaxa}
\alias{routine_planktoscope_ecotaxa}
\title{routine_planktoscope_ecotaxa}
\usage{
routine_planktoscope_ecotaxa()
}
\value{
Biovolumes, dataframes and resume tables, figures to check the data.
}
\description{
A routine to compute biovolumes and save updated csv files. Save also table resume for each taxonomic group and figures to check the data.
1. Choose a directory and create a new one inside to save results.
2. Use function check_metadata to allow metadata edit.
3. Compute biovolume and BSS tables with chosen metadata.
4. Save results and resume tables (AB or BV for each taxa by sample).
5. Graphical output : metadata, project and samples.
}
\examples{
routine_planktoscope_ecotaxa()
}
