% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_helper-functions.R
\name{timeOfDay}
\alias{timeOfDay}
\title{Calculate the time of day as day, twilight, buffer}
\usage{
timeOfDay(time, sunrise, sunset, buffer, tz = "UTC")
}
\arguments{
\item{time}{a character vector of times (formatted to be from parmeta)}

\item{sunrise}{in UVP timezone (typically UTC)}

\item{sunset}{sunset}

\item{buffer}{range around sunset/sunrise to assign as twilight in hours}

\item{tz}{timezone}
}
\description{
Calculate the time of day as day, twilight, buffer
}
\examples{
timeOfDay("14:00:00", "06:00:00", "18:00:00", buffer = .5) #Should return day

}
\author{
Alex Barth
}
