% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopart_trim-obj.R
\name{trim_par}
\alias{trim_par}
\title{Trim par_list}
\usage{
trim_par(par_list, trim_cat, trim_amt, operator)
}
\arguments{
\item{par_list}{a par_file list}

\item{trim_cat}{the category to trim by}

\item{trim_amt}{the amount to trim by}

\item{operator}{either 'less','greater','lessequal', or 'greaterequal'}
}
\description{
A simple function to trim par list based on some criteria
}
