% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopart_zoo_conc.R
\name{uvp_conc}
\alias{uvp_conc}
\title{Calculate UVP Concentration from Ecopart Object}
\usage{
uvp_conc(ecopart_obj, cast_name, depth_breaks, ...)
}
\arguments{
\item{ecopart_obj}{an ecopart object list}

\item{cast_name}{name of the cast to get concentration of}

\item{depth_breaks}{a vector of depth break ranges}

\item{...}{pass to \code{\link[=bin_taxa]{bin_taxa()}} for func and func_col}
}
\description{
This function can take in an ecopart object (from ecopart_import)
And for a given cast name, it will return the concentration.
Relies on default formatting and will return count concentration by default
However, you can use func_col and func arguements to pass to bin_taxa
if trying to get biomass/biovolume concentration.
}
\author{
Alex Barth
}
