% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopart_zoo_conc.R
\name{uvp_zoo_conc}
\alias{uvp_zoo_conc}
\title{Calculate UVP Zooplankton Concentration from an Ecopart Object}
\usage{
uvp_zoo_conc(ecopart_obj, cast_name = NULL, breaks, ...)
}
\arguments{
\item{ecopart_obj}{an ecopart object list}

\item{cast_name}{the name (or names of a cast)}

\item{breaks}{a vector to break on}

\item{...}{pass into \code{\link[=bin_taxa]{bin_taxa()}}}
}
\description{
This function takes in an ecopart object and will calculate the
concentration for zooplankton data. This is a big wrapper function to
build and add functionality to \code{\link[=uvp_conc]{uvp_conc()}}.
}
