% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zooregroup_zooscan}
\alias{zooregroup_zooscan}
\title{zooregroup_zooscan}
\format{
## `zooregroup_zooscan`
}
\usage{
zooregroup_zooscan
}
\description{
A database with trophic levels and OTU.
}
\keyword{datasets}
